+++ Sample/src/main/java/com/angcyo/dsltablayout/demo/SegmentFragment.kt
@@ -86,6 +86,7 @@ class SegmentFragment : BaseTabLayoutFragment() {
                             tabBadge.gradientSolidColor = randomColor()
 
                             tabBadge.updateOriginDrawable()
+                            null
                         }
                     }
                 }